+++ app/src/main/java/widget/missions/normal/MissionWidgetNormal.kt
@@ -40,12 +40,12 @@ import data.getImageFromAfxId
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import tools.bitmapResize
-import tools.createMissionCircularProgressBarBitmap
-import tools.getAsset
-import tools.getMissionDurationRemaining
-import tools.getMissionPercentComplete
-import tools.getShipName
+import tools.utilities.bitmapResize
+import tools.utilities.createMissionCircularProgressBarBitmap
+import tools.utilities.getAsset
+import tools.utilities.getMissionDurationRemaining
+import tools.utilities.getMissionPercentComplete
+import tools.utilities.getShipName
 import widget.missions.MissionWidgetDataStore
 import widget.missions.MissionWidgetDataStorePreferencesKeys
 import widget.missions.MissionWidgetUpdater
